Change virtual class allocation methods and add documentation.
Previously, portable modules were implemented as virtual classes. This
meant that instantiation had to be done via base class pointers, and no
local instances were possible. To make things more flexible for the user,
the portable modules are no longer virtual and instead own a pointer to
an implementation class that can be extended to perform specialized code.