Reduce project scope
Originally ULTRA240 was designed as a minimal framework, but that made
integrating other libraries difficult. Rather than the user integrating
other libraries into ULTRA240, ULTRA240 should be easily integratable
into other libraries and frameworks. This is a major rewrite of the
project and big step towards that goal.

The windowing, input detection, and system-independent file system
abstraction components have been removed. What remains is the renderer,
resource loader, dynamic library abstraction, and entity engine.

The renderer has been redesigned to that it can be dropped into an
existing pipeline. How easily that is remains to be seen.